+++ src/link/core/index.ts
@@ -1,11 +1,8 @@
-export {
-  empty,
-  from,
-  split,
-  concat,
-  execute,
-  LinkError,
-  ApolloLink,
-} from './link';
+export { empty } from './empty';
+export { from } from './from';
+export { split } from './split';
+export { concat } from './concat';
+export { execute } from './execute';
+export { ApolloLink } from './ApolloLink';
 
 export * from './types';
